feat: Remove colorMode option
