+++ packages/client/src/elements/HTMLInspectElement/index.css
@@ -1,4 +1,10 @@
 :host {
+  --text-color: #ffffff;
+  --text-color2: #dddddd;
+  --bg-color: #2c2c2e;
+  --bg-color-opt: #2c2c2ed9;
+  --bg-color2: #6c6c6e;
+  --cyan: #4df9fa;
   --red: #ff335c;
   --red-light: #ff335c33;
   --shadow: 0 0 1px var(--bg-color2);
