+++ packages/client/src/elements/HTMLToggleElement/index.css
@@ -27,10 +27,9 @@
 }
 .oe-active {
   color: var(--cyan);
-  box-shadow: 0 0 30px 6px var(--cyan);
 }
 .oe-touch {
-  padding: 8px;
+  padding: 10px;
 }
 .oe-touch .oe-button {
   padding: 3px;
